+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 - widgets
   - frames: allow loading local html files without requiring a web server to serve them
   - text: escape html tags characters (`<` & `>`)
-
+  - input: added `asYouType` property
+  
 ## 0.47.4
 
 - bug fixes